+++ src/Service/MediaConvert/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - AWS api-change: This release adds support for AVC passthrough, the ability to specify PTS offset without padding, and an A/V segment matching feature.
+- AWS api-change: This release adds a configurable Quality Level setting for the top rendition of Auto ABR jobs
 
 ## 1.6.0